requirements.txt
# -----------------------------------------------------------------------
#           Dependencies needed to run the app
# -----------------------------------------------------------------------

psycopg2
config
snowflake-connector-python[pandas]
streamlit
pandas
plotly